refactor: 移除外部 scrcpy 依赖，统一使用内部 StreamPlayer
- 废弃 controller/scrcpy.rs（外部进程管理）
- 废弃 app/ui/player.rs（V4L2Player）
- 移除初始化流程中的 Scrcpy 启动逻辑
- 修复所有 Clippy 警告和 Doctest 错误
- 代码减少 314 行

架构变化：
  旧：外部 scrcpy 进程 → V4L2 → V4L2Player
  新：内部 scrcpy 协议 → StreamPlayer（VAAPI/NVDEC）

测试：34/34 通过
格式检查：通过
Clippy：通过（0 警告）

+++ src/decoder/nvdec.rs
@@ -152,8 +152,8 @@ impl NvdecDecoder {
                     // Extract NV12 data
                     let y_plane = sw_frame.data(0);
                     let uv_plane = sw_frame.data(1);
-                    let y_linesize = sw_frame.stride(0) as usize;
-                    let uv_linesize = sw_frame.stride(1) as usize;
+                    let y_linesize = sw_frame.stride(0);
+                    let uv_linesize = sw_frame.stride(1);
 
                     info!(
                         "NV12 frame layout: {}x{}, Y linesize={} UV linesize={}",
